TT-680: Add types to createResponseHandler
Rename helpers.ts as createResponseHandler.ts in order to better indicate what
the file contains.

Add types to passport-strategy functions and createResponseHandler
callbacks so that typescript would guide the user to implement the right
thing.

Remove unnecessary Symbol from USER_NOT_ACCEPTED_ERROR as a normal
string is sufficient, more usable and conforms to the errorhandling
function types.

remote yarn from pre-commit.sh

Author: @tunylund
