+++ lib/index.ts
@@ -1,4 +1,4 @@
 import { createStrategy } from './passport-verify-strategy'
-import { createResponseHandler } from './helpers'
+import { createResponseHandler } from './create-response-handler'
 
 export { createStrategy, createResponseHandler }
